refactor: rename package and remove onComplete callback
DATA1-2833

- Renamed package to `@dataone/embedded-connection-journey`
- Removed the onComplete callback as it is no longer required
- Updated README.md to reflect changes
