+++ README.md
@@ -4,11 +4,11 @@ Public packages that support partners integrating with D·One services. This mon
 
 ## 🚀 Packages
 
-### [@dataone/iframe-connection-journey](./packages/iframe-connection-journey/)
+### [@dataone/embedded-connection-journey](./packages/embedded-connection-journey/)
 
 A lightweight JavaScript library that enables you to seamlessly integrate the D·One connection journey into your website using an iframe overlay. Perfect for partners who want to embed the D·One open banking connection flow without redirecting users away from their site.
 
-[📖 View Documentation](./packages/iframe-connection-journey/README.md) | [📦 NPM Package](https://www.npmjs.com/package/@dataone/iframe-connection-journey)
+[📖 View Documentation](./packages/embedded-connection-journey/README.md) | [📦 NPM Package](https://www.npmjs.com/package/@dataone/embedded-connection-journey)
 
 ## 🔗 Links
