+++ packages/embedded-connection-journey/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@dataone/iframe-connection-journey",
+  "name": "@dataone/embedded-connection-journey",
   "version": "0.1.0",
   "description": "Launch the D·One connection journey in an iframe within your own website",
   "main": "dist/esm/index.min.mjs",
@@ -18,7 +18,7 @@
   "repository": {
     "type": "git",
     "url": "https://github.com/clearscore/data-one-integrations.git",
-    "directory": "packages/iframe-connection-journey"
+    "directory": "packages/embedded-connection-journey"
   },
   "keywords": [
     "data-one",
