+++ packages/embedded-connection-journey/vite.config.ts
